refactor(search-types): relocate PredicateShortfall to foundation crate (TD-FLIGHT-1 step 1)
Move the PredicateShortfall diagnostic struct (TD-064) from the modality-tier
proximadb-observability-engine into the foundation proximadb-search-types crate
so transport-neutral search DTOs can reference it without an upward dependency.
The definition is re-exported from its historical search_plan_trace module path,
so every existing crate::observability::search_plan_trace::PredicateShortfall
caller keeps compiling unchanged.

Prerequisite for routing Arrow Flight search through the canonical v2
RecordSearchPort (TD-FLIGHT-1).
