+++ crates/modalities/proximadb-observability-engine/Cargo.toml
@@ -61,6 +61,10 @@ proximadb-data-model.workspace = true
 proximadb-kernel.workspace = true
 proximadb-records.workspace = true
 proximadb-tenant.workspace = true
+# Foundation: owns `PredicateShortfall`, which the search-plan-trace module
+# re-exports (TD-FLIGHT-1 relocated the definition here so transport-neutral
+# search DTOs can reference it without a modality-tier dependency).
+proximadb-search-types.workspace = true
 # Optional, feature-gated (otlp-metering) — the OTLP metering *export* surface.
 # Same specs as the root crate's optional deps.
 opentelemetry = { version = "0.32", optional = true }
